+++ CHANGELOG.md
@@ -28,8 +28,10 @@
 ## [Unreleased][]
 ### Added
 ### Changed
+- UI: duplicated background submenu.
 ### Removed
 ### Fixed
+- UI: duplicated background submenu.
 - UI: refresh rate submenu not showing the actual setting used (with a check).
 
 ## [2.4.0] - 2024-07-24